Develop

+++ src/main/java/com/example/ormi5projectteam4/repository/AnimalRepository.java
@@ -1,7 +1,7 @@
 package com.example.ormi5projectteam4.repository;
 
-import com.example.ormi5projectteam4.domain.Animal;
+import com.example.ormi5projectteam4.domain.entity.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalRepository extends JpaRepository<Animal, Integer> {
+public interface AnimalRepository extends JpaRepository<Animal, Long> {
 }
