+++ src/main/java/com/example/ormi5projectteam4/repository/ImageRepository.java
@@ -1,7 +1,7 @@
 package com.example.ormi5projectteam4.repository;
 
-import com.example.ormi5projectteam4.domain.Image;
+import com.example.ormi5projectteam4.domain.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
+public interface ImageRepository extends JpaRepository<Image, Long> {
 }
